requirements.txt
streamlit==1.36.0
Pillow==10.3.0
streamlit-image-select==0.6.0
google-generativeai==0.7.1
python-dotenv==1.0.1
weave